requirements.txt
sphinx
nose
spam-blocklists
whois
backports.ssl_match_hostname
python-dateutil
